+++ packages/react/src/components/Table/Table.story.jsx
@@ -5,7 +5,7 @@ import Arrow from '@carbon/icons-react/es/arrow--right/16';
 import Add from '@carbon/icons-react/es/add/16';
 import Edit from '@carbon/icons-react/es/edit/16';
 import { spacing03 } from '@carbon/layout';
-import { Add20, Column20, TrashCan16 } from '@carbon/icons-react';
+import { Add20, Column20, TrashCan16, ViewOff16 } from '@carbon/icons-react';
 import { cloneDeep, assign, isEqual } from 'lodash-es';
 import { firstBy } from 'thenby';
 import uuid from 'uuid';
@@ -422,6 +422,7 @@ export const tableActions = {
     // since they won't be adding this prop to any of their components
     // can be readded in V3.
     // onToggleAggregations: action('onToggleAggregations'),
+    onApplyToolbarAction: action('onApplyToolbarAction'),
   },
   table: {
     onRowClicked: action('onRowClicked'),
@@ -559,6 +560,30 @@ export const initialState = {
   },
 };
 
+const tableToolbarActions = [
+  {
+    id: 'edit',
+    labelText: 'Edit',
+    renderIcon: 'edit',
+    disabled: true,
+    isOverflow: true,
+  },
+  {
+    id: 'delete',
+    labelText: 'Delete',
+    isDelete: true,
+    hasDivider: true,
+    isOverflow: true,
+    renderIcon: () => <TrashCan16 />,
+  },
+  {
+    id: 'hidden',
+    labelText: 'Hidden',
+    hidden: true,
+    isOverflow: true,
+  },
+];
+
 export default {
   title: '1 - Watson IoT/Table/Table',
 
@@ -691,6 +716,7 @@ export const BasicDumbTable = () => {
         toolbar: {
           activeBar: hasColumnSelection || hasColumnSelectionConfig ? 'column' : undefined,
           isDisabled: boolean('Disable the table toolbar (view.toolbar.isDisabled)', false),
+          toolbarActions: tableToolbarActions,
         },
         table: {
           loadingState: {
@@ -2029,6 +2055,14 @@ export const WithFilters = () => {
         },
         toolbar: {
           activeBar: 'filter',
+          toolbarActions: [
+            ...tableToolbarActions,
+            {
+              id: 'toggle',
+              labelText: 'toolbarAction shown in toolbar instead of overflow',
+              renderIcon: ViewOff16,
+            },
+          ],
           customToolbarContent: (
             <div style={{ alignItems: 'center', display: 'flex', padding: '0 1rem' }}>
               custom content
@@ -2043,7 +2077,7 @@ export const WithFilters = () => {
   );
 };
 
-WithFilters.storyName = 'with filtering and custom toolbar content';
+WithFilters.storyName = 'with filtering, toolbarActions, and custom toolbar content';
 
 export const WithAdvancedFilters = () => {
   const operands = {